Sec-Financial-Data-Pipeline/backend/requirements.txt
fastapi
uvicorn
snowflake-connector-python
pandas
google-cloud-secret-manager
python-dotenv
pydantic